+++ latticejson/__init__.py
@@ -1,3 +1,4 @@
 from .io import read, convert
-from .validate import validate, UndefinedObjectError
+from .validate import validate
 from .convert import elegant_to_latticejson, latticejson_to_elegant
+from .exceptions import UndefinedObjectError
