+++ tests/test_validate.py
@@ -1,15 +1,15 @@
-import os
+from pathlib import Path
 import pytest
 from latticejson.validate import validate_file
 from latticejson import UndefinedObjectError
 
-dir_name = os.path.dirname(__file__)
+base_dir = Path(__file__).resolve().parent / "data"
 
 
 def test_validate():
-    file_path = os.path.join(dir_name, "data", "fodo.json")
+    file_path = base_dir / "fodo.json"
     validate_file(file_path)
 
-    file_path = os.path.join(dir_name, "data", "test_undefined.json")
+    file_path = base_dir / "test_undefined.json"
     with pytest.raises(UndefinedObjectError):
         validate_file(file_path)
